LC-minimum-number-of-days-to-make-m-bouquets [Runtime: N/A; Memory: N/A]

+++ array/minimum-number-of-days-to-make-m-bouquets.cpp
@@ -7,7 +7,7 @@ public:
             mi=min(arr[i],mi);
         }
         int mid;
-        if(arr.size()<m*k){
+        if(arr.size()<(long long)m*k){
             return -1;
         }
         else{
